Move platforms code from pkg/install to pkg/platform
As they should be moved there long time ago, when we introduced Platform
interface.
