+++ cli/cmd/root.go
@@ -21,9 +21,9 @@ import (
 	"github.com/spf13/viper"
 
 	// Register platforms by adding an anonymous import.
-	_ "github.com/kinvolk/lokomotive/pkg/install/aws"
-	_ "github.com/kinvolk/lokomotive/pkg/install/baremetal"
-	_ "github.com/kinvolk/lokomotive/pkg/install/packet"
+	_ "github.com/kinvolk/lokomotive/pkg/platform/aws"
+	_ "github.com/kinvolk/lokomotive/pkg/platform/baremetal"
+	_ "github.com/kinvolk/lokomotive/pkg/platform/packet"
 
 	// Register backends by adding an anonymous import.
 	_ "github.com/kinvolk/lokomotive/pkg/backend/local"
