+++ pkg/platform/aws/template.go
