+++ pkg/platform/baremetal/baremetal.go
