+++ pkg/platform/baremetal/template.go
