+++ pkg/platform/packet/template.go
